# build/openapi/definition.json
{
  "openapi": "3.0.3",
...
  "info": {
    "title": "Bank Api",
    "description": "Bank Api",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "{protocol}://{environment}/{version}",
      "variables": {
        "protocol": {
          "default": "http",
          "enum": [
            "http",
            "https"
          ]
        },
        "environment": {
          "default": "localhost:8080",
          "enum": [
            "localhost:8080"
          ]
        },
        "version": {
          "default": "v1",
          "enum": [
            "v1"
          ]
        }
      }
    }
  ],
  "paths": {
...
    "/docs/definition": {
      "get": {
        "tags": [
          "documentation"
        ],
        "summary": "Openapi definition",
        "description": "Retrieve the openapi definition",
        "operationId": "definition",
        "responses": {
          "200": {
            "description": "Openapi definition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/UnexpectedErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
...
      "UnexpectedErrorResponse": {
        "description": "Unexpected error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GenericError"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "GenericError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string",
                "nullable": true
              },
              "details": {
                "type": "string",
                "nullable": true
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}